+++ src/users/users.service.ts
@@ -7,12 +7,14 @@ import {
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Users } from './entity/user.entity';
+import { Users } from './entity/users.entity';
 import { CreateUserDto } from './dtos/create-user.dto';
 import { UpdateUserDto } from './dtos/update-user.dto';
-import { Organizations } from '../organizations/entity/organization.entity';
+import { Organizations } from '../organizations/entity/organizations.entity';
 import * as bcrypt from 'bcrypt';
 import { ValidateUser } from './dtos/validate-user.dto';
+import { Request } from 'express';
+import { User } from './interfaces/user.interface';
 
 @Injectable()
 export class UsersService {
@@ -32,7 +34,7 @@ export class UsersService {
       });
 
       if (existingUser) {
-        throw new ConflictException('Email already exists');
+        throw new ConflictException('Email already exists in the organization');
       }
 
       const organization = await this.organizationsRepository.findOne({
@@ -67,11 +69,16 @@ export class UsersService {
     }
   }
 
-  async findAllUsers(): Promise<Users[]> {
+  async findAllUsers(request: Request): Promise<Users[]> {
     try {
+      const organizationId = (request?.user as User)?.organizationId;
       const users = await this.usersRepository.find({
+        where: { organization: { id: organizationId } },
         relations: ['organization'],
       });
+      if (!users) {
+        throw new NotFoundException('User not found');
+      }
       return users;
     } catch (error) {
       this.logger.error(
@@ -85,17 +92,16 @@ export class UsersService {
     }
   }
 
-  async findOneUser(id: string): Promise<Users> {
+  async findOneUser(id: string, request: Request): Promise<Users> {
     try {
+      const organizationId = (request?.user as User)?.organizationId;
       const user = await this.usersRepository.findOne({
-        where: { id },
+        where: { id, organization: { id: organizationId } },
         relations: ['organization'],
       });
-
       if (!user) {
         throw new NotFoundException('User not found');
       }
-
       return user;
     } catch (error) {
       this.logger.error(
@@ -112,9 +118,13 @@ export class UsersService {
   async updateOneUser(
     id: string,
     updateUserDto: UpdateUserDto,
+    request: Request,
   ): Promise<Users> {
     try {
-      const user = await this.usersRepository.findOne({ where: { id } });
+      const organizationId = (request?.user as User)?.organizationId;
+      const user = await this.usersRepository.findOne({
+        where: { id, organization: { id: organizationId } },
+      });
 
       if (!user) {
         throw new NotFoundException('User not found');
@@ -136,15 +146,51 @@ export class UsersService {
     }
   }
 
-  async deleteOneUser(id: string): Promise<void> {
+  async updateUserLastLoginTimestamp(
+    id: string,
+    lastLoginTimestamp: number,
+  ): Promise<Users> {
     try {
-      const user = await this.usersRepository.findOne({ where: { id } });
+      const user = await this.usersRepository.findOne({
+        where: { id },
+      });
 
       if (!user) {
         throw new NotFoundException('User not found');
       }
+      Object.assign(user, { lastLoginTimestamp });
+      await this.usersRepository.save(user);
+
+      return user;
+    } catch (error) {
+      this.logger.error(
+        JSON.stringify({
+          context: 'Update User',
+          resource: UsersService.name,
+          message: error.message,
+        }),
+      );
+      throw error;
+    }
+  }
+
+  async deleteOneUser(
+    id: string,
+    request: Request,
+  ): Promise<{ message: string }> {
+    try {
+      const userId = (request?.user as User)?.id;
+      const organizationId = (request?.user as User)?.organizationId;
+      const user = await this.usersRepository.findOne({
+        where: { id, organization: { id: organizationId } },
+      });
+
+      if (userId != id || !user) {
+        throw new NotFoundException('User not found');
+      }
 
       await this.usersRepository.remove(user);
+      return { message: 'Record deleted successfully' };
     } catch (error) {
       this.logger.error(
         JSON.stringify({
@@ -163,6 +209,7 @@ export class UsersService {
       const user = await this.usersRepository.findOne({
         where: { email },
         relations: ['organization'],
+        select: { id: true, password: true },
       });
 
       if (!user || !(await bcrypt.compare(password, user.password))) {
